% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIALOGUE.util.R
\name{call.plot.multilabels}
\alias{call.plot.multilabels}
\title{call.plot.multilabels}
\usage{
call.plot.multilabels(
  X,
  labels,
  main = NULL,
  xlab = "UMAP1",
  ylab = "UMAP2",
  add.N = F,
  pch = 16,
  cex = 0.3,
  cex.axis = 0.6,
  set.flag = F
)
}
\description{
call.plot.multilabels
}
