% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIALOGUE.util.R
\name{call.plot.plus}
\alias{call.plot.plus}
\title{call.plot.plus}
\usage{
call.plot.plus(
  x,
  y = NULL,
  labels,
  b.top,
  red.top = F,
  regression.flag = F,
  my.col = NULL,
  set.flag = F,
  cor.flag = F,
  pch = 16,
  cex = 0.3,
  main = "",
  ylab = "tSNE2",
  xlab = "tSNE1",
  cex.axis = 0.6,
  add.N = F,
  grey.zeros = F,
  legend.flag = T
)
}
\description{
call.plot.plus
}
