% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIALOGUE.util.R
\name{call.plot}
\alias{call.plot}
\title{call.plot}
\usage{
call.plot(
  x,
  y = NULL,
  labels,
  regression.flag = F,
  my.col = NULL,
  set.flag = F,
  cor.flag = F,
  legend.flag = T,
  pch = 16,
  cex = 0.5,
  main = "",
  ylab = "UMAP2",
  xlab = "UMAP1",
  cex.axis = 0.6,
  add.N = F,
  cex.main = 1,
  color.spec = "rgb"
)
}
\description{
call.plot
}
