% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIALOGUE.cell.type.R
\docType{class}
\name{cell.type-class}
\alias{cell.type-class}
\alias{cell.type}
\title{DIALOGUE cell.type S4 class}
\description{
An S4 class that represents a subset of cells (for example, a particular cell type).
}
\section{Slots}{

\describe{
\item{\code{name}}{cell type name;}

\item{\code{cells}}{cell identifiers (1xn);}

\item{\code{tpm}}{gene expression matrix (mxn)}

\item{\code{genes}}{genes (1xm) represented in the [tpm] matrix;}

\item{\code{X}}{features matrix (kxn), e.g., PCs, NMF components, tpm etc.}

\item{\code{samples}}{the samples corresponding to the cells in [cells] (1xn)}
}}

\seealso{
See \href{https://github.com/livnatje/DIALOGUE}{DIALOGUE GitHub page} for more details.
\code{\link{DIALOGUE.plot}}
}
\author{
Livnat Jerby-Arnon
}
