% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIALOGUE.main.R
\name{DIALOGUE.feature_select}
\alias{DIALOGUE.feature_select}
\title{DIALOGUE.feature_select}
\usage{
DIALOGUE.feature_select(rA, config)
}
\arguments{
\item{rA}{list of \linkS4class{cell.type} objects.}

\item{config}{configuration parameters}
}
\value{
A list with the following components:

quality controlled latent features most informative
}
\description{
Selects latent features
}
\author{
Livnat Jerby-Arnon, modified by David Banh
}
