% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIALOGUE.main.R
\name{DIALOGUE.mcp}
\alias{DIALOGUE.mcp}
\title{DIALOGUE.mcp}
\usage{
DIALOGUE.mcp(rA, config)
}
\arguments{
\item{rA}{list of \linkS4class{cell.type} objects.}

\item{config}{configuration parameters}
}
\value{
A list with the following components:

R list of \linkS4class{cell.type} objects.
}
\description{
Runs a hierarchical linear model to identify multicellular programs of coordinated gene sets
}
\author{
Livnat Jerby-Arnon, modified by David Banh
}
