% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIALOGUE.plot.R
\name{DIALOGUE.plot}
\alias{DIALOGUE.plot}
\title{DIALOGUE.plot}
\usage{
DIALOGUE.plot(R, results.dir = "~/Desktop/DIALOGUE.results/", pheno = NULL)
}
\arguments{
\item{rA}{list of cell type objects (see cell.type)}
}
\description{
Plot DIALOGUE results.
The resulting plots will show the cell-type-specific components
of each multicellular program as a function of its other components;
and the composition of each multicellular program, depicting how many of its genes
are cell-type-specific, and how many are shared across multiple cell types.
}
\examples{
R<-DIALOGUE.run(rA,results.dir = "~/Desktop/Results/")
R<-DIALOGE.plot(R,results.dir = "~/Desktop/Figures/")
Alternatively
R<-DIALOGUE.run(rA,results.dir = "~/Desktop/Results/",plot.flag = T)
}
\author{
Livnat Jerby-Arnon
}
