% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIALOGUE.main.R
\name{DIALOGUE.run_PMD}
\alias{DIALOGUE.run_PMD}
\title{DIALOGUE.run_PMD}
\usage{
DIALOGUE.run_PMD(rA, X, config = config)
}
\arguments{
\item{rA}{list of \linkS4class{cell.type} objects.}

\item{X}{list of latent components per \linkS4class{cell.type} object.}

\item{config}{configuration parameters}
}
\value{
A list with the following components:

R list of \linkS4class{cell.type} objects.
}
\description{
Runs penalised matrix decomposition on multiple matrices via Multiple Canonical Correlation Analysis
}
\author{
Livnat Jerby-Arnon, modified by David Banh
}
