% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIALOGUE.main.R
\name{DIALOGUE.summary}
\alias{DIALOGUE.summary}
\title{DIALOGUE.summary}
\usage{
DIALOGUE.summary(rA, config)
}
\arguments{
\item{rA}{list of \linkS4class{cell.type} objects.}

\item{config}{configuration parameters}
}
\value{
A list with the following components:

R list of \linkS4class{cell.type} objects.
}
\description{
Summarises previous feature selection, PMD, and MCP steps
}
\author{
Livnat Jerby-Arnon, modified by David Banh
}
