% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIALOGUE_SeuratExample.R
\name{DIALOGUE_make.cell.type.seurat}
\alias{DIALOGUE_make.cell.type.seurat}
\title{DIALOGUE_make.cell.type.seurat}
\usage{
DIALOGUE_make.cell.type.seurat(obj, cell.subtypes, name)
}
\arguments{
\item{obj}{Seurat object}

\item{cell.subtypes}{the cell type or subtype to be use. Make sure the obj metadata includes a column of "cell.subtypes"}

\item{name}{the name to be used in the DIALOGUE run to refer to this cell type.}
}
\description{
Using Seurat objects as input for DIALOGUE make.cell.type function
}
\examples{
# Run DIALOGUE with simulated PBMC data
r1<-DIALOGUE_make.cell.type.seurat(obj, cell.subtypes = "CD8.T.cell", name = "CD8.T.cell")
r1<- DIALOGUE_make.cell.type.seurat(obj, cell.subtypes = c("Naive CD4 T","Memory CD4 T","CD8 T"), name = "T.cell")
r2<- DIALOGUE_make.cell.type.seurat(obj, cell.subtypes = c("CD14+ Mono","FCGR3A+ Mono","DC"), name = "Myeloid")
rA<- list(A = r1,B = r2)
R <- DIALOGUE.run(rA = rA, main = "ToyExample",k = 2, results.dir = "DIALOGUE.output",spatial.flag = F,plot.flag = T,conf = "cellQ")

}
