% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIALOGUE_SeuratExample.R
\name{DIALOGUE_SeuratExample}
\alias{DIALOGUE_SeuratExample}
\title{DIALOGUE_SeuratExample}
\usage{
DIALOGUE_SeuratExample(results.dir)
}
\arguments{
\item{results.dir}{path to the directory for saving the output.}
}
\description{
Using Seurat objects as input for DIALOGUE
}
\examples{
# Run DIALOGUE with simulated PBMC data
R<-DIALOGUE_SeuratExample(results.dir = "DIALOGUE.output")

}
