% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIALOGUE.main.R
\name{DLG.get.param}
\alias{DLG.get.param}
\title{DLG.get.param}
\usage{
DLG.get.param(
  k = 2,
  results.dir = DLG.file("/Results/MCPs/"),
  seed1 = 1234,
  plot.flag = F,
  penalty = NULL,
  MCP.cell.types = NULL,
  frm = NULL,
  pheno = NULL,
  PMD2 = F,
  single.BH = F,
  extra.sparse = F,
  abn.c = 15,
  conf = "cellQ",
  covar = c("cellQ", "tme.qc"),
  n.genes = 200,
  averaging.function = colMedians,
  p.anova = 0.05,
  find.genes = T,
  specific.pair = NULL,
  center.flag = T,
  bypass.emp = F,
  parallel.vs = F,
  spatial.flag = F,
  full.version = T
)
}
\arguments{
\item{k}{the number of multicellular programs to identify;}

\item{results.dir}{path to the results directory, where the output will be saved;}

\item{plot.flag}{if TRUE then \code{\link{DIALOGUE.plot}} will be called to plot the results; default is FALSE;}

\item{abn.c}{the minimal number of cells that a sample should have to be considered for MCP detection;}

\item{spatial.flag}{should be TRUE if working with spatial data with small niches, and TRUE if working with single cell data or larger tissue microenvironment niches. The default value is FALSE.}
}
\value{
A list with the parameters as input for DIALOGUE.run
}
\description{
DIALOGUE is a dimensionality reduction approach that uses cross-cell-type
associations to identify multicellular programs (MCPs) and map the cell transcriptome
as a function of its environment.
}
\seealso{
See \href{https://github.com/livnatje/DIALOGUE}{DIALOGUE GitHub page} for more details.
\code{\link{DIALOGUE.plot}}
}
\author{
Livnat Jerby
}
