% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIALOGUE.cell.type.R
\name{make.cell.type}
\alias{make.cell.type}
\title{make.cell.type}
\usage{
make.cell.type(
  name,
  tpm,
  samples,
  X = NULL,
  metadata = NULL,
  tpmAv = t(average.mat.rows(t(tpm), samples)),
  cellQ
)
}
\arguments{
\item{name}{cell type name}

\item{tpm}{gene expression or any type of single-cell profiling (mxn)}

\item{samples}{the sample of each cell (1xn)}

\item{X}{features matrix (kxn), e.g., PCs, NMF components, tpm etc.; these features will be used to identify the multicellular programs.}

\item{metadata}{cell features (rxn)}

\item{cellQ}{cell quality measures, e.g., number of reads/genes detected (1xn)}
}
\description{
This function generates a \linkS4class{cell.type} object for DIALOGUE.
}
\section{Fields}{

\describe{
\item{\code{cell.type}}{a representation of a specific type of cells}
}}

