% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIALOGUE.cell.type.R
\docType{class}
\name{cell.type-class}
\alias{cell.type-class}
\alias{cell.type}
\title{DIALOGUE cell.type S4 class}
\description{
An S4 class that represents a subset of cells (for example, a particular cell type).
}
\section{Slots}{

\describe{
\item{\code{name}}{cell type name;}

\item{\code{cells}}{cell identifiers (n x 1);}

\item{\code{tpm}}{gene expression matrix (m x n)}

\item{\code{genes}}{genes (m x 1) represented in the [tpm] matrix;}

\item{\code{X}}{features matrix (n x k), where n is the number of cells and k is the number of features, e.g., PCs, NMF components, tpm etc.}

\item{\code{samples}}{the samples corresponding to the cells in [cells] (n x 1)}
}}

\seealso{
See \href{https://github.com/livnatje/DIALOGUE}{DIALOGUE GitHub page} for more details.
\code{\link{DIALOGUE.plot}}
}
\author{
Livnat Jerby-Arnon
}
