% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIALOGUE.main.R
\name{DIALOGUE.run}
\alias{DIALOGUE.run}
\title{DIALOGUE.run}
\usage{
DIALOGUE.run(rA, main, param, plot.flag = T)
}
\arguments{
\item{rA}{list of \linkS4class{cell.type} objects.}

\item{main}{the name of the run that will be used for naming the output files in the results directory}

\item{param}{the parameters of the run - these will also be saved with the output for reproducibility.
See \code{\link{DLG.get.param}} for more information.}

\item{plot.flag}{if TRUE then \code{\link{DIALOGUE.plot}} will be called to plot the results; default is FALSE;}
}
\value{
A list with the following components:

sig -  the multicellular programs, given as a list of signatures;

scores - the multicellular programs' scores in each cell;

gene.pval - the cross-cell-type p-values of each program;

pref - the correlation (R) and association (mixed-effects p-value) between the cell-type-sepcific
components of each multicellular programs.
}
\description{
DIALOGUE is a dimensionality reduction approach that uses cross-cell-type
associations to identify multicellular programs (MCPs) and map the cell transcriptome
as a function of its environment.
}
\seealso{
See \href{https://github.com/livnatje/DIALOGUE}{DIALOGUE GitHub page} for more details.
\code{\link{DIALOGUE.plot}}
}
\author{
Livnat Jerby
}
