% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIALOGUE.util.R
\name{dialogue_config}
\alias{dialogue_config}
\title{dialogue_config}
\usage{
dialogue_config(verbose = F)
}
\arguments{
\item{verbose}{Print out the configuration parameters?}
}
\value{
configuration parameters
}
\description{
Summarises previous feature selection, PMD, and MCP steps
}
\author{
Livnat Jerby-Arnon, modified by David Banh
}
